# zespolowe/calc_ser_git/calculator_server/swagger_server/models/number.py
# coding: utf-8

from __future__ import absolute_import
from datetime import date, datetime  # noqa: F401

from typing import List, Dict  # noqa: F401

from swagger_server.models.base_model_ import Model
from swagger_server import util


class Number(Model):
    """NOTE: This class is auto generated by the swagger code generator program.

    Do not edit the class manually.
    """
    def __init__(self, value: int=None):  # noqa: E501
        """Number - a model defined in Swagger

        :param value: The value of this Number.  # noqa: E501
        :type value: int
        """
        self.swagger_types = {
            'value': int
        }

        self.attribute_map = {
            'value': 'value'
        }
        self._value = value

    @classmethod
    def from_dict(cls, dikt) -> 'Number':
        """Returns the dict as a model

        :param dikt: A dict.
        :type: dict
        :return: The Number of this Number.  # noqa: E501
        :rtype: Number
        """
        return util.deserialize_model(dikt, cls)

    @property
    def value(self) -> int:
        """Gets the value of this Number.


        :return: The value of this Number.
        :rtype: int
        """
        return self._value

    @value.setter
    def value(self, value: int):
        """Sets the value of this Number.


        :param value: The value of this Number.
        :type value: int
        """

        self._value = value